==> 5-reproducibility.Rmd <==
---
title: "SPARS trial A vs B"
subtitle: "Reproducibility of the stimulus-response relationship"
author: "Peter Kamerman"
date: "`r format(Sys.Date(), '%d %b %Y')`"
output: 
  html_document:
    keep_md: true
    theme: yeti
    hightlight: pygments
    toc: true
    toc_depth: 5
    toc_float: true
    code_folding: show
---

```{r setup, include = FALSE}
# Load packages
library(magrittr)
library(tidyverse)
library(lqmm)

# Set ggplot theme (since I've loaded sjPlot)
sjPlot::set_theme(base = theme_bw())

# Colour palette
cb_palette <- c('#F0E442', '#0072B2', '#D55E00', '#CC79A7')

# knitr setup
knitr::opts_chunk$set(warning = FALSE,
                      message = FALSE,
                      fig.retina = 2,
                      fig.height = 7,
                      fig.width = 7,
                      fig.align = 'center',
                      fig.path = 'figures/5-reproducibility/')
```

----

This analysis examines the _'reproducibility'_ of the stimulus-response characateristics of the SPARS. We assessed reproducibility by examining what proportion of SPARS ratings recorded during Trial B fell within the 95% prediction interval of the cubic model of the stimulus-response relationship described in Trial A. 

Please note that the modelling process for Trial A (and hence the calculation of the 95% predictuion interval of the model) used Tukey trimeans of the raw SPARS rating generated by each participant at each stimulus intensity. Also stimulus intensities used in Trial A ranged from 1.00 to 4.00J (and were the same for each participant), whereas the stimulus intensities used in Trial B ranged from 1.75 to 4.50J (and could differ between participants). Therefore, for some participants in Trial B, the upper range of the stimulus intensities they were exposed to lay outside of the range used to calculate the 95% prediction interval for Trial A. These 'outlying' values were not included in the analysis.

We assessed _'reproducibility'_ under two conditions:

1.  Raw SPARS ratings collected for each participant in Trial B at each stimulus intensity (each participant had 8 exposures to each of the stimulus intensities they were tested at) were assessed for inclusion in the 95% prediction interval.

2. Tukey trimeans calculated from of SPARS ratings for each participant in Trial B at each stimulus intensity were assessed for inclusion in the 95% prediction interval.

Condition 1 assesses whether once-off recordings fall with a prediction interval calculated using data 'averaged' across multiple recordings. Condition 2 assesses whether data generated in a similar manner to that used to calculate the prediction interval (using 'averaged' data) fall within the interval. 

----

# Import and inspect data
```{r import}
# Import and inspect
## Trial A
data_A <- read_rds('./data-cleaned/SPARS_A.rds')
glimpse(data_A)

## Trial B
data_B <- read_rds('./data-cleaned/SPARS_B.rds')
glimpse(data_B)
```

----

# Clean data

```{r clean}
############################################################
#                                                          #
#                 Define tri.mean function                 #
#                                                          #
############################################################
tri.mean <- function(x) {
  # Calculate quantiles
  q1 <- quantile(x, probs = 0.25, na.rm = TRUE)[[1]]
  q2 <- median(x, na.rm = TRUE)
  q3 <- quantile(x, probs = 0.75, na.rm = TRUE)[[1]]
  # Calculate trimean
  tm <- (q2 + ((q1 + q3) / 2)) / 2
  # Convert to integer
  tm <- as.integer(round(tm))
  return(tm)
}

############################################################
#                                                          #
#                       Trial A data                       #
#                                                          #
############################################################
data_tmA <- data_A %>%
    # Select columns
    select(PID, intensity, rating) %>%
    # Calculate tri.mean
    group_by(PID, intensity) %>% 
    summarise(tri_mean = tri.mean(rating))
    
############################################################
#                                                          #
#                       Trial B data                       #
#                                                          #
############################################################
data_B %<>%
    # Only retain SPARS data
    filter(scale == 'SPARS')

data_tmB <- data_B %>% 
    # Select columns
    select(PID, intensity, rating) %>% 
    # Calculate tri.mean
    group_by(PID, intensity) %>% 
    summarise(tri_mean = tri.mean(rating))
```

----

# Trial A 95% prediction interval vs 'raw' Trial B comparison

### Exploratory plots

```{r explore_predRAW, fig.width = 9}
############################################################
#                                                          #
#                    Trial A quantiles                     #
#                                                          #
############################################################
# Quantile model with 2.5, 25, 50, 75, and 97.5% quantiles
qmm <- lqmm(fixed = tri_mean ~ poly(intensity, 3),
            random = ~ intensity,
            group = PID,
            data = data_tmA,
            tau = c(0.025, 0.975))

# Get predicted values
## Level 0 (conditional, note difference to the lmer diagnostics)
quant_predict <- as.data.frame(predict(qmm, level = 0))
names(quant_predict) <- paste0('Q', c(2.5, 97.5))

# Join with 'data_tmA'
data_tmA %<>%
  bind_cols(quant_predict)

# Trim prediction to upper and lower limits of the scale
data_tmA %<>%
  mutate_if(is.numeric,
            funs(ifelse(. > 50,
                        yes = 50,
                        no = ifelse(. < -50,
                                    yes = -50,
                                    no = .))))

############################################################
#                                                          #
#             Trial B within Trial A quantiles             #
#                                                          #
############################################################
# Process data
raw_predictionPlots <- data_B %>%
    group_by(PID) %>%
    nest() %>%
    # Plot data
    mutate(plots = map2(.x = data,
                        .y = PID,
                       ~ ggplot() +
                           geom_ribbon(data = data_tmA,
                                       aes(x = intensity,
                                           ymin = `Q2.5`,
                                           ymax = `Q97.5`),
                                       fill = '#CCCCCC') +
                           geom_point(data = .x,
                                      aes(x = intensity,
                                          y = rating),
                                      colour = cb_palette[[2]]) +
                           geom_smooth(data = .x,
                                      aes(x = intensity,
                                          y = rating),
                                      colour = cb_palette[[2]],
                                      method = 'loess',
                                      se = FALSE) +
                           geom_hline(yintercept = 0,
                                      linetype = 2) +
                           labs(title = paste0(.y, ': Raw SPARS ratings vs stimulus intensity from Trial B, facetted by experimental block'),
                                subtitle = 'Blue circles: raw data points | Blue lines: loess curve | Grey area: 95% prediction interval of the Trial A dataset\nThe prediction interval extends to the limits of the range of stimulus intensities used in Trial A (1 to 4J)',
                                x = 'Stimulus intensity (J)',
                                y = 'SPARS rating [-50 to 50]') +
                           scale_y_continuous(limits = c(-50, 50)) +
                           scale_x_continuous(limits = c(1, 5)) +
                           facet_wrap(~block_number, ncol = 2))) 

# Print output
walk(.x = raw_predictionPlots$plots, ~print(.x))
```

Participant 5 (ID05) has very odd data, with a dog-leg shaped response curve and virtually all the data falling outside of the prediction interval. 

### Quantification

```{r quant_predRAW}
# Get Trial A prediction interval bounds for each stimulus intensity 
pred_bounds <- data_tmA %>%
    ungroup() %>%
    select(intensity, Q2.5, Q97.5) %>%
    # `base::unique` giving 15 instead of 13 rows, so move to plan B
    group_by(intensity) %>%
    summarise(Q2.5 = mean(Q2.5),
              Q97.5 = mean(Q97.5))

# View prediction interval bounds
knitr::kable(pred_bounds,
             caption = '95% prediction interval bounds for Trial A (cubic model) at each stimulus intensity', 
             col.names = c('Stimulus intensity', 'Q2.5', 'Q97.5'), 
             digits = 2,
             align = 'lrr')

# Filter Trial B data to the stimulus range used in Trial A (1 to 4J)
data_BReduceRAW <- data_B %>% 
    ungroup() %>% 
    filter(intensity >= 1 & intensity <= 4)

# Calculate whether raw points are inside or outside the 
# prediction interval at each stimulus intensity
predict_RAW <- data_BReduceRAW %>% 
    # Nest data by stimulus intensity
    group_by(intensity) %>%
    nest() %>%
    arrange(intensity) %>%
    # Join with pred_bounds
    left_join(pred_bounds) %>% 
    # Perform calculation 
    mutate(analysis = pmap(.l = list(data, Q2.5, Q97.5),
                           ~ mutate(.data = ..1,
                                    included = case_when(
                                        rating >= ..2 & rating <= ..3 ~ 'yes',
                                        TRUE ~ 'no'
                                    )))) %>%
    # Drop columns and bring it all back together
    select(-data, -(starts_with('Q'))) %>%
    unnest()
    
# Plot
predict_RAW %>%
    # Calculate counts
    group_by(intensity, included) %>% 
    summarise(count = n()) %>%
    ungroup() %>% 
    mutate(intensity = sprintf('%.2f', intensity)) %>%
    # Plot
    ggplot(data = .) +
    aes(y = count,
        x = intensity,
        fill = included) +
    geom_bar(stat = 'identity', 
             position = position_fill()) +
    labs(title = 'Proportion of raw SPARS ratings falling within the 95% prediction interval of the Trial A (cubic model)',
         x = 'Stimulus intensity (J)',
         y = 'Proportion of ratings') +
    scale_fill_manual(name = 'Included in interval: ',
                      values = rev(cb_palette[2:3])) +
    theme(legend.position = 'top')
```

### Quantification after omitting ID05

```{r quant_predRAW2}
# Plot
predict_RAW %>%
    # Filter out ID05
    filter(PID != 'ID05') %>%
    # Calculate counts
    group_by(intensity, included) %>% 
    summarise(count = n()) %>%
    ungroup() %>% 
    mutate(intensity = sprintf('%.2f', intensity)) %>%
    # Plot
    ggplot(data = .) +
    aes(y = count,
        x = intensity,
        fill = included) +
    geom_bar(stat = 'identity', 
             position = position_fill()) +
    labs(title = 'Proportion of raw SPARS ratings falling within the 95% prediction interval of the Trial A (cubic model)',
         subtitle = '[WITH ID05 REMOVED]',
         x = 'Stimulus intensity (J)',
         y = 'Proportion of ratings') +
    scale_fill_manual(name = 'Included in interval: ',
                      values = rev(cb_palette[2:3])) +
    theme(legend.position = 'top')
```

The proportion of values falling outside the prediction interval falls after removing ID05, but the pattern of the relationship between stimulus intensity and inclusion within the prediction interval does not change much. 

----

# Trial A 95% prediction interval vs 'Tukey trimean' Trial B comparison

### Exploratory plots

```{r plot_predTM, fig.width = 9}
# Process data
tm_predictionPlots <- data_tmB %>%
    group_by(PID) %>%
    nest() %>%
    # Plot data
    mutate(plots = map2(.x = data,
                        .y = PID,
                       ~ ggplot() +
                           geom_ribbon(data = data_tmA,
                                       aes(x = intensity,
                                           ymin = `Q2.5`,
                                           ymax = `Q97.5`),
                                       fill = '#CCCCCC') +
                           geom_point(data = .x,
                                      aes(x = intensity,
                                          y = tri_mean),
                                      colour = cb_palette[[2]]) +
                           geom_smooth(data = .x,
                                      aes(x = intensity,
                                          y = tri_mean),
                                      colour = cb_palette[[2]],
                                      method = 'loess',
                                      se = FALSE) +
                           geom_hline(yintercept = 0,
                                      linetype = 2) +
                           labs(title = paste0(.y, ': Tukey trimeans of raw SPARS ratings vs stimulus intensity from Trial B, facetted by experimental block'),
                                subtitle = 'Blue circles: Tukey trimean data points | Blue lines: loess curve | Grey area: 95% prediction interval of the Trial A dataset\nThe prediction interval extends to the limits of the range of stimulus intensities used in Trial A (1 to 4J)',
                                x = 'Stimulus intensity (J)',
                                y = 'SPARS rating [-50 to 50]') +
                           scale_y_continuous(limits = c(-50, 50)) +
                           scale_x_continuous(limits = c(1, 5)))) 

# Print output
walk(.x = tm_predictionPlots$plots, ~print(.x))
```

As expected the raw data plots, participant 5 (ID05) has very odd data, with a dog-leg shaped response curve and **all** the Tukey trimean values falling outside of the prediction interval. 

### Quantification

```{r quant_predTM}
# Filter Trial B data to the stimulus range used in Trial A (1 to 4J)
data_BReduceTM <- data_tmB %>% 
    ungroup() %>% 
    filter(intensity >= 1 & intensity <= 4)

# Calculate whether raw points are inside or outside the 
# prediction interval at each stimulus intensity
predict_TM <- data_BReduceTM %>% 
    # Nest data by stimulus intensity
    group_by(intensity) %>%
    nest() %>%
    arrange(intensity) %>%
    # Join with pred_bounds
    left_join(pred_bounds) %>% 
    # Perform calculation 
    mutate(analysis = pmap(.l = list(data, Q2.5, Q97.5),
                           ~ mutate(.data = ..1,
                                    included = case_when(
                                        tri_mean >= ..2 & tri_mean <= ..3 ~ 'yes',
                                        TRUE ~ 'no'
                                    )))) %>%
    # Drop columns and bring it all back together
    select(-data, -(starts_with('Q'))) %>%
    unnest()
    
# Plot
predict_TM %>%
    # Calculate counts
    group_by(intensity, included) %>% 
    summarise(count = n()) %>%
    ungroup() %>% 
    mutate(intensity = sprintf('%.2f', intensity)) %>%
    # Plot
    ggplot(data = .) +
    aes(y = count,
        x = intensity,
        fill = included) +
    geom_bar(stat = 'identity', 
             position = position_fill()) +
    labs(title = 'Proportion of Tukey trimeans of raw SPARS ratings falling within the 95% prediction interval of the Trial A (cubic model)',
         x = 'Stimulus intensity (J)',
         y = 'Proportion of ratings') +
    scale_fill_manual(name = 'Included in interval: ',
                      values = rev(cb_palette[2:3])) +
    theme(legend.position = 'top')
```

### Quantification without ID05

```{r quant_predTM2}
# Plot
predict_TM %>%
    # Filter out ID05
    filter(PID != 'ID05') %>%
    # Calculate counts
    group_by(intensity, included) %>% 
    summarise(count = n()) %>%
    ungroup() %>% 
    mutate(intensity = sprintf('%.2f', intensity)) %>%
    # Plot
    ggplot(data = .) +
    aes(y = count,
        x = intensity,
        fill = included) +
    geom_bar(stat = 'identity', 
             position = position_fill()) +
    labs(title = 'Proportion of Tukey trimeans of raw SPARS ratings falling within the 95% prediction interval of the Trial A (cubic model)',
         subtitle = '[WITH ID05 REMOVED]',
         x = 'Stimulus intensity (J)',
         y = 'Proportion of ratings') +
    scale_fill_manual(name = 'Included in interval: ',
                      values = rev(cb_palette[2:3])) +
    theme(legend.position = 'top')
```

When using Tukey trimean data, omitting ID05 results in a substantial change improvement in the proportion of data falling within the prediction interval. However, there are still issues in the stimulus intensity range: 2.25 to 3.00J.

----

# Session information
```{r session_info}
sessionInfo()
```
